requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.3.3
google-api-python-client>=2.33.0
python-dotenv>=0.19.2
requests>=2.26.0
spotipy>=2.19.0
ytmusicapi>=1.5.0
sqlalchemy>=1.4.0
alembic>=1.7.0
pytest>=7.0.0
pytest-cov>=4.0.0
